requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
dj-database-url==0.5.0
Django==4.0.2
django-cors-headers==3.13.0
django-online-users==0.3
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dotenv==0.19.2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.8
whitenoise==6.2.0
